3/4 there

+++ LinkedLists/QuizLinkedList.java
@@ -5,7 +5,7 @@ public class QuizLinkedList extends GradeLinkedList {
     public QuizLinkedList() {
         super();
     }
-    public QuizLinkedList(int points, int totalPoints, String name) {
+    public QuizLinkedList(double points, double totalPoints, String name) {
         super(points, totalPoints, name);
     }
